added container image and added it as a atextre

+++ OpenGLEngine/Triangle.h
@@ -22,12 +22,14 @@ public:
 	{
 		material->use();
 
+
 		int tintLocation = glGetUniformLocation(
 			material->shaderProgram, "tintColor");
 		glUniform4f(tintLocation, red, green, blue, 1.0);
 
-		float offsetLocation = glGetUniformLocation(material->shaderProgram, "offset");
+		int offsetLocation = glGetUniformLocation(material->shaderProgram, "offset");
 		glUniform2f(offsetLocation, offsetX, offsetY);
+
 
 		mesh->Render();
 	}
